+++ client/laundroclean/src/app/admin/(without-dashboard-layout)/controlpanel/layout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ControlPanelLayoutComponent from "src/components/ui/Modals/CompanyUser/Admin/ControlPanel/ControlPanelLayout";
+import ControlPanelLayoutComponent from "src/components/layouts/CompanyUser/Admin/ControlPanelLayoutUI";
 
 export default async function ControlPanelLayout ({children} : {children: React.ReactNode}) {
     return (
